Relocate task-update functionality in a new component
