+++ src/app/worklists/editor/task-form-editor.component.scss
@@ -10,13 +10,3 @@
   height: 600px;
   width: 720px;
 }
-
-.form-frame {
-  height: 430px;
-  overflow-y: auto;
-}
-
-.form {
-  width: 640px;
-  overflow-y: auto;
-}
