+++ src/app/worklists/editor/task-update.component.html
@@ -1,0 +1,79 @@
+<table class="form">
+  <tr>
+    <td class="form-title form-title-text" colspan="2">Información general de la tarea o actividad</td>
+  </tr>
+  <tr>
+    <td>Nombre:</td>
+    <td> <input type="text" class="text-box" [(ngModel)]="selectedTask.name"></td>
+  </tr>
+  <tr>
+    <td>Observaciones:</td>
+    <td><textarea class="text-area" rows="5" cols="25" [(ngModel)]="selectedTask.notes"></textarea></td>
+  </tr>
+</table>
+<table class="form">
+  <tr>
+    <td class="form-title form-title-text" colspan="3">Responsable de la tarea y porcentaje de progreso</td>
+  </tr>
+  <tr>
+    <td>Responsable:</td>
+    <td>
+      <select class="select-box" style="width:340px" [(ngModel)]="selectedTask.responsibleUID">
+          <option value="">( Seleccionar )</option>
+          <option value="Undefined">No asignado</option>
+          <option *ngFor="let responsible of responsiblesList"
+                  [value]="responsible.uid">{{responsible.name}}</option>
+        </select>
+    </td>
+    <td class="text-align-right">
+      Progreso: &nbsp; &nbsp;
+      <select class="select-box" style="width:60px" [(ngModel)]="selectedTask.progress">
+          <option value="0">0%</option>
+          <option value="10">10%</option>
+          <option value="20">20%</option>
+          <option value="25">25%</option>
+          <option value="30">30%</option>
+          <option value="33">33%</option>
+          <option value="40">40%</option>
+          <option value="50">50%</option>
+          <option value="60">60%</option>
+          <option value="66">66%</option>
+          <option value="70">70%</option>
+          <option value="75">75%</option>
+          <option value="80">80%</option>
+          <option value="90">90%</option>
+          <option value="95">95%</option>
+          <option value="100">100%</option>
+        </select>
+    </td>
+  </tr>
+</table>
+<table class="form">
+  <tr>
+    <td class="form-title form-title-text" colspan="3">Fecha objetivo y fecha legal</td>
+  </tr>
+  <tr>
+    <td>Fecha objetivo:</td>
+    <td style="width:70%">
+      <input type="text" id="calendar" class="text-box" style="width:100px" [ngModel]="selectedTask.targetDate | date: 'dd-MM-yyyy'" (target)="selectedTask=$value">
+      <button id="calendarButton">..</button>
+    </td>
+    <td class="text-align-right">Fecha legal: &nbsp; <b>{{task.dueDate | date: 'dd-MMM-yyyy'}}</b></td>
+  </tr>
+
+</table>
+<div class="form-footer">
+  <table class="form">
+    <tr>
+      <td style='width:33%'>
+        <button class="btn" (click)="cancel()">Cancelar</button>
+      </td>
+      <td *ngIf="selectedTask.stage !== 'Done'" class="text-align-center" style='width:33%'>
+        <button class="btn-warning" (click)="onCloseTask()">Concluir la tarea</button>
+      </td>
+      <td class="text-align-right" style='width:33%'>
+        <button class="btn-action" (click)="onUpdateTask()">Guardar los cambios</button>
+      </td>
+    </tr>
+  </table>
+</div>
