+++ src/app/worklists/editor/task-update.component.scss
@@ -1,0 +1,11 @@
+@import "../../../styles/editor-window-style";
+@import "../../../styles/form";
+.form-frame {
+  height: 430px;
+  overflow-y: auto;
+}
+
+.form {
+  width: 640px;
+  overflow-y: auto;
+}
